# docs/frontleaves_yggleaf_swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员分页查询用户列表，支持角色、关键词、时间范围筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-用户接口"
                ],
                "summary": "[超管] 用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页条数(最大100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色筛选(SUPER_ADMIN/ADMIN/PLAYER)",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词搜索(用户名/邮箱)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "注册时间起始(RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "注册时间截止(RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.AdminUserListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "需要超级管理员权限",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/admin/users/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户完整详情，包含账户信息、游戏档案配额、资源库配额及皮肤/披风资源列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-用户接口"
                ],
                "summary": "[超管] 用户详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标用户 ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.AdminUserDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "需要超级管理员权限",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/user/game-password": {
            "put": {
                "description": "已通过 OAuth2 认证的用户可直接设置/重置游戏密码，无需旧密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[玩家] 更新游戏密码",
                "parameters": [
                    {
                        "description": "更新游戏密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateGamePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserCurrentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "根据 AT 获取用户信息，获取到本程序的用户信息（含账户完善状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[玩家] 用户信息",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserCurrentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求体格式不正确",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "用户已禁用或账户已锁定",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "admin.AdminCapeItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "texture_url": {
                    "type": "string"
                }
            }
        },
        "admin.AdminSkinItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "model": {
                    "description": "\"STEVE\" 或 \"ALEX\"",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "texture_url": {
                    "type": "string"
                }
            }
        },
        "admin.AdminUserBasic": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "has_ban": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "jailed_at": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "admin.AdminUserDetailResponse": {
            "type": "object",
            "properties": {
                "cape_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.AdminCapeItem"
                    }
                },
                "game_profile": {
                    "$ref": "#/definitions/admin.GameProfileQuotaInfo"
                },
                "library_quota": {
                    "$ref": "#/definitions/admin.LibraryQuotaInfo"
                },
                "skin_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.AdminSkinItem"
                    }
                },
                "user": {
                    "$ref": "#/definitions/admin.AdminUserBasic"
                }
            }
        },
        "admin.AdminUserItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "has_ban": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "admin.AdminUserListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.AdminUserItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "admin.GameProfileQuotaInfo": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "used": {
                    "type": "integer"
                }
            }
        },
        "admin.LibraryQuotaInfo": {
            "type": "object",
            "properties": {
                "capes_private_total": {
                    "type": "integer"
                },
                "capes_private_used": {
                    "type": "integer"
                },
                "capes_public_total": {
                    "type": "integer"
                },
                "capes_public_used": {
                    "type": "integer"
                },
                "skins_private_total": {
                    "type": "integer"
                },
                "skins_private_used": {
                    "type": "integer"
                },
                "skins_public_total": {
                    "type": "integer"
                },
                "skins_public_used": {
                    "type": "integer"
                }
            }
        },
...
        "entity.CapeLibrary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "description": "是否公开",
                    "type": "boolean"
                },
                "name": {
                    "description": "披风名称",
                    "type": "string"
                },
                "texture": {
                    "description": "披风纹理文件ID(雪花算法)",
                    "type": "integer"
                },
                "texture_hash": {
                    "description": "披风纹理哈希",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "----------\n 外键约束\n----------",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "创建者/上传者用户ID(为空代表系统内置披风)",
                    "type": "integer"
                }
            }
        },
        "entity.GameProfile": {
            "type": "object",
            "properties": {
                "cape_library": {
                    "description": "关联披风库",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.CapeLibrary"
                        }
                    ]
                },
                "cape_library_id": {
                    "description": "关联披风库ID",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "游戏内用户名",
                    "type": "string"
                },
                "skin_library": {
                    "description": "关联皮肤库",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.SkinLibrary"
                        }
                    ]
                },
                "skin_library_id": {
                    "description": "关联皮肤库ID",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "----------\n 外键约束\n----------",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "关联用户ID",
                    "type": "integer"
                },
                "uuid": {
                    "description": "Minecraft UUID",
                    "type": "string"
                }
            }
        },
...
        "entity.LibraryQuota": {
            "type": "object",
            "properties": {
                "capes_private_total": {
                    "description": "披风配额",
                    "type": "integer"
                },
                "capes_private_used": {
                    "description": "私有披风已使用额度",
                    "type": "integer"
                },
                "capes_public_total": {
                    "description": "公开披风总额度",
                    "type": "integer"
                },
                "capes_public_used": {
                    "description": "公开披风已使用额度",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "skins_private_total": {
                    "description": "皮肤配额",
                    "type": "integer"
                },
                "skins_private_used": {
                    "description": "私有皮肤已使用额度",
                    "type": "integer"
                },
                "skins_public_total": {
                    "description": "公开皮肤总额度",
                    "type": "integer"
                },
                "skins_public_used": {
                    "description": "公开皮肤已使用额度",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "----------\n 外键约束\n----------",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "关联用户ID",
                    "type": "integer"
                }
            }
        },
...
        "entity.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "display_name": {
                    "description": "角色显示名称",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.RoleName"
                        }
                    ]
                }
            }
        },
        "entity.RoleName": {
            "type": "string",
            "enum": [
                "SUPER_ADMIN",
                "ADMIN",
                "PLAYER"
            ],
            "x-enum-comments": {
                "RoleAdmin": "管理员角色名称",
                "RolePlayer": "普通玩家角色名称",
                "RoleSuperAdmin": "超级管理员角色名称"
            },
            "x-enum-descriptions": [
                "超级管理员角色名称",
                "管理员角色名称",
                "普通玩家角色名称"
            ],
            "x-enum-varnames": [
                "RoleSuperAdmin",
                "RoleAdmin",
                "RolePlayer"
            ]
        },
...
        "entity.User": {
            "type": "object",
            "properties": {
                "cape_libraries": {
                    "description": "披风库关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CapeLibrary"
                    }
                },
                "email": {
                    "description": "用户邮箱",
                    "type": "string"
                },
                "game_profiles": {
                    "description": "游戏档案关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.GameProfile"
                    }
                },
                "has_ban": {
                    "description": "用户是否被封禁禁止登录",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "jailed_at": {
                    "description": "用户被监禁的时间",
                    "type": "string"
                },
                "library_quotas": {
                    "description": "资源库配额关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.LibraryQuota"
                    }
                },
                "phone": {
                    "description": "用户手机号",
                    "type": "string"
                },
                "role": {
                    "description": "----------\n 外键约束\n----------",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Role"
                        }
                    ]
                },
                "role_name": {
                    "description": "关联角色名称",
                    "type": "string"
                },
                "skin_libraries": {
                    "description": "皮肤库关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.SkinLibrary"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_cape_libraries": {
                    "description": "用户披风关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UserCapeLibrary"
                    }
                },
                "user_skin_libraries": {
                    "description": "用户皮肤关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.UserSkinLibrary"
                    }
                },
                "username": {
                    "description": "用户用户名",
                    "type": "string"
                }
            }
        },
...
        "user.UpdateGamePasswordRequest": {
            "type": "object",
            "required": [
                "confirm_password",
                "new_password"
            ],
            "properties": {
                "confirm_password": {
                    "description": "确认新密码（需与 new_password 一致）",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                },
                "new_password": {
                    "description": "新游戏密码（6-128 字符）",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                }
            }
        },
        "user.UserCurrentResponse": {
            "type": "object",
            "properties": {
                "extend": {
                    "description": "扩展信息（含账户完善状态）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/user.UserExtend"
                        }
                    ]
                },
                "user": {
                    "description": "用户实体信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.User"
                        }
                    ]
                }
            }
        },
        "user.UserExtend": {
            "type": "object",
            "properties": {
                "account_ready": {
                    "description": "AccountReady 账户完善状态。\n\n  - \"ready\":          所有必要信息已填写完毕\n  - \"game_password\":  游戏密码未设置（返回缺失字段名）",
                    "type": "string"
                }
            }
        },
        "xBase.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
    }
}